% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{addResource,easyreporting-method}
\alias{addResource,easyreporting-method}
\title{addResource}
\usage{
\S4method{addResource}{easyreporting}(object, source, reference, description)
}
\arguments{
\item{object}{an easyreporting class instance}

\item{source}{a string indicating the reference (i.e. "GEO")}

\item{reference}{a string indicanting the reference of the source (i.e. "GSE60231)}

\item{description}{a natural language description}
}
\value{
an easyreporting class instance
}
\description{
addResource
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
rd <- addResource(rd, source="GEO", "GSE60231", "Transcriptome of BMDC to different antigen delivery systems")
\dontrun{
compile(rd)
}
}
