% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{compile,easyreporting-method}
\alias{compile,easyreporting-method}
\title{compile}
\usage{
\S4method{compile}{easyreporting}(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
none
}
\description{
prints the sessionInfo and compiles the rmarkdown file
}
\examples{
\dontrun{
rd <- easyreporting(filenamePath="./project_report", title="example_report",
                        author=c("It's me"))
compile(rd)
}
}
