% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6MarkdownMethods.R
\name{compile}
\alias{compile}
\title{compile}
\usage{
compile()
}
\value{
none
}
\description{
compiles the rmarkdown file
}
\examples{
rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
rd$compile()
}
