% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingC.R
\name{easyreporting}
\alias{easyreporting}
\alias{easyreporting-class}
\title{easyreporting}
\usage{
easyreporting(filenamePath, title, author, optionList = NULL)

easyreporting(filenamePath, title, author, optionList = NULL)
}
\arguments{
\item{filenamePath}{the path with the name of the rmarkdown.}

\item{title}{the title of the report section.}

\item{author}{the author of the report.}

\item{optionList}{a list of options for the general rmarkdown document.}
}
\value{
an S4 easyreporting class instance
}
\description{
An S4 class for managing rmarkdown report.
Each instance describes an rmarkdown file.
}
\section{Slots}{

\describe{
\item{\code{filenamePath}}{the path with the name of the rmarkdown.}

\item{\code{title}}{the title of the report section.}

\item{\code{author}}{the author of the report.}

\item{\code{documentType}}{the type of the final document (fixed to "html").}

\item{\code{optionList}}{a list of options for the general rmarkdown document.}
}}

\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author="It's me")
}
