% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6MarkdownClass.R
\docType{class}
\name{easyreporting}
\alias{easyreporting}
\title{easyreporting}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
easyreporting(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
}
\arguments{
\item{filenamepath}{the path with the name of the rmarkdown.}

\item{title}{the title fo the report.}

\item{author}{the author of the report.}
}
\value{
an easyreporting R6 class instance
}
\description{
An R6 class for managing rmarkdown report.
Each instance describes an rmarkdown file.
}
\examples{
rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
}
\keyword{datasets}
