% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{erGUIVolcano}
\alias{erGUIVolcano}
\title{erGUIVolcano}
\usage{
erGUIVolcano()
}
\value{
none
}
\description{
executes a Shiny GUI to perform a volcano plot and trace its
executed functions.
}
\examples{
\dontrun{
erGUIVolcano()
}
}
