% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{getBibliography,easyreporting-method}
\alias{getBibliography,easyreporting-method}
\title{getBibliography}
\usage{
\S4method{getBibliography}{easyreporting}(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
none
}
\description{
returns the bibfile name attached to the object
}
\examples{
\dontrun{
# TBD
}
}
