% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getBibliography}
\alias{getBibliography}
\title{getBibliography}
\usage{
getBibliography(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
none
}
\description{
returns the bibfile name attached to the object
}
\examples{
\dontrun{
TBD
}
}
