% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsM.R
\name{getOptionsList,easyreporting-method}
\alias{getOptionsList,easyreporting-method}
\title{getOptionsList}
\usage{
\S4method{getOptionsList}{easyreporting}(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
a list of options
}
\description{
returns the optionList from the easyreporting class
}
\examples{
\dontrun{
rd <- easyreporting(filenamePath="./project_report", title="example_report",
                        author=c("It's me"))
optList <- getOptionsList(rd)
}
}
