% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{getOptionsList}
\alias{getOptionsList}
\title{getOptionsList}
\usage{
getOptionsList()
}
\value{
a list of options
}
\description{
returns the optionsList from the easyreporting class
}
\examples{
rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
optList <- rd$getOptionsList()

}
