% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{getReportFilename,easyreporting-method}
\alias{getReportFilename,easyreporting-method}
\title{getReportFilename}
\usage{
\S4method{getReportFilename}{easyreporting}(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
a string of report file name with path
}
\description{
returns the report filename with path
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
(rep <- getReportFilename(rd))
}
