% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{getReportFilename}
\alias{getReportFilename}
\title{getReportFilename-generic}
\usage{
getReportFilename(object)
}
\arguments{
\item{object}{a general object}
}
\value{
none
}
\description{
getReportFilename-generic
}
\keyword{internal}
