% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{initReportFilename}
\alias{initReportFilename}
\title{initReportFilename-generic}
\usage{
initReportFilename(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
initReportFilename-generic
}
\keyword{internal}
