% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6MarkdownMethods.R
\name{makeOptionsList}
\alias{makeOptionsList}
\title{maketOptionsList}
\usage{
makeOptionsList(cacheFlag = TRUE, evalFlag = TRUE, echoFlag = TRUE,
  warningFlag = FALSE, showMessages = FALSE, includeFlag = TRUE)
}
\arguments{
\item{cacheFlag}{boolean for caching chunk data (default TRUE)}

\item{evalFlag}{boolean for evaluating the code chunk in the compiled version
(default TRUE)}

\item{echoFlag}{boolean for showing the code chunk (default TRUE)}

\item{warningFlag}{boolean for showing the chunk warnings (default FALSE)}

\item{showMessages}{boolean for showing the chunk warnings in compiled
version (default FALSE)}

\item{includeFlag}{boolean for including the code chunk in the compiled
version (default TRUE)}
}
\value{
list of rmarkdown options
}
\description{
makes an list of rmarkdown options
}
\examples{
optList <- makeOptionsList()
}
