% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdCodeChunkCommented}
\alias{mkdCodeChunkCommented}
\title{mkdCodeChunkCommented-generic}
\usage{
mkdCodeChunkCommented(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
mkdCodeChunkCommented-generic
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
optList <- makeOptionsList(includeFlag=TRUE, cacheFlag=TRUE)
mkdCodeChunkCommented(rd,
                commentMsg="This is the comment of the following code chunk",
                codeMsg="a <- 1\nb <- 2\n(c <- a+b)\n", optionList=optList,
                sourceFilesList=NULL)

}
