% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdCodeChunkEnd,easyreporting-method}
\alias{mkdCodeChunkEnd,easyreporting-method}
\title{mkdCodeChunkEnd}
\usage{
\S4method{mkdCodeChunkEnd}{easyreporting}(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
none
}
\description{
it creates a code chunk end. Always use it after a
mkdCodeChunkSt()
}
\examples{
 rd <- easyreporting(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
mkdCodeChunkSt(rd)
## just leaving empty
mkdCodeChunkEnd(rd)
}
