% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6MarkdownMethods.R
\name{mkdCodeChunkSt}
\alias{mkdCodeChunkSt}
\title{mkdCodeChunkSt}
\usage{
mkdCodeChunkSt(optionsList = self$getOptionsList(),
  source.files.list = NULL)
}
\arguments{
\item{optionsList}{a list of options}

\item{source.files.list}{a list of files that can be sourced inside the code
chunk.}
}
\value{
none
}
\description{
it creates a code chunk start. A list of options and files to
source  for the chunk can optionally be passed to the function.
}
\examples{
 rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
## no options
rd$mkdCodeChunkSt()
## just leaving empty
rd$mkdCodeChunkEnd()

## setting options
optList <- makeOptionsList(includeFlag=TRUE)
rd$mkdCodeChunkSt(optionsList=optList)
## just leaving empty
rd$mkdCodeChunkEnd()

}
