% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing.R
\name{mkdCodeChunkTitledCommented,easyreporting-method}
\alias{mkdCodeChunkTitledCommented,easyreporting-method}
\title{mkdCodeChunkTitledCommented}
\usage{
\S4method{mkdCodeChunkTitledCommented}{easyreporting}(
  object,
  title = NULL,
  level = 1,
  comment = NULL,
  code,
  optionList = getOptionsList(object),
  sourceFilesList = NULL
)
}
\arguments{
\item{object}{an easyreporting class object}

\item{title}{the title to assign to the code chunk section}

\item{level}{the level of the title (default is 1)}

\item{comment}{a string with the natural language comment for the chunk.}

\item{code}{a string within the code.}

\item{optionList}{a list of options (default is the class options).}

\item{sourceFilesList}{a optional list of files to source inside the chunk.}
}
\value{
none
}
\description{
it creates a complete code chunk, adding a natural language
comment before of it.
}
\examples{
\dontrun{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
optList <- makeOptionsList(includeFlag=TRUE, cacheFlag=TRUE)
mkdCodeChunkTitledCommented(rd, title="Title Example", level=1,
                comment="This is the comment of the following code chunk",
                code="a <- 1\n b <- 2\n(c <- a+b)\n", optionList=optList,
                sourceFilesList=NULL)
}
}
