% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdCodeChunkTitledCommented}
\alias{mkdCodeChunkTitledCommented}
\title{mkdCodeChunkCommented}
\usage{
mkdCodeChunkTitledCommented(
  object,
  title = NULL,
  level = 1,
  commentMsg = NULL,
  codeMsg,
  optionList = getOptionsList(object),
  sourceFilesList = NULL
)
}
\arguments{
\item{object}{an easyreporting class object}

\item{title}{the title to assign to the code chunk section}

\item{level}{the level of the title (default is 1)}

\item{commentMsg}{a string with the natural language comment for the chunk.}

\item{codeMsg}{a string within the code.}

\item{optionList}{a list of options (default is the class options).}

\item{sourceFilesList}{a optional list of files to source inside the chunk.}
}
\value{
none
}
\description{
it creates a complete code chunk, adding a natural language
comment before of it.
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
optList <- makeOptionsList(includeFlag=TRUE, cacheFlag=TRUE)
mkdCodeChunkCommented(rd,
                commentMsg="This is the comment of the following code chunk",
                codeMsg="a <- 1\nb <- 2\n(c <- a+b)\n", optionList=optList,
                sourceFilesList=NULL)

}
