% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdGeneralMsg,easyreporting-method}
\alias{mkdGeneralMsg,easyreporting-method}
\title{mkdGeneralMsg}
\usage{
\S4method{mkdGeneralMsg}{easyreporting}(object, message)
}
\arguments{
\item{object}{an easyreporting class object}

\item{message}{the message to append to the report}
}
\value{
none
}
\description{
It appends a general message to the report.
Useful for adding natural language comments.
}
\examples{
\dontrun{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
mkdGeneralMsg(rd, "Writing a paragraph to describe my code chunk")
}
}
