% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdGeneralMsg}
\alias{mkdGeneralMsg}
\title{mkdGeneralMsg-generic}
\usage{
mkdGeneralMsg(object, ...)

mkdGeneralMsg(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none

none
}
\description{
mkdGeneralMsg-generic

mkdGeneralMsg-generic
}
\keyword{internal}
