% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdGeneralTitledMsg,easyreporting-method}
\alias{mkdGeneralTitledMsg,easyreporting-method}
\title{mkdGeneralTitledMsg}
\usage{
\S4method{mkdGeneralTitledMsg}{easyreporting}(object, title = NULL, level = 1, message)
}
\arguments{
\item{object}{an easyreporting class object}

\item{title}{the optional title to give to the message section}

\item{level}{the level (1 to 6) of the title (default is 1)}

\item{message}{the message to append to the report}
}
\value{
none
}
\description{
It appends a a titled section followed by a general message to 
the report. Useful for adding natural language comments.
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
mkdGeneralMsg(rd, title="Generic SubTitle for this message", level=2,
message="Writing a paragraph to describe my code chunk")
}
