% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsM.R
\name{mkdSetGlobalOpts,easyreporting-method}
\alias{mkdSetGlobalOpts,easyreporting-method}
\title{mkdSetGlobalOpts}
\usage{
\S4method{mkdSetGlobalOpts}{easyreporting}(object, optionList = list())
}
\arguments{
\item{object}{an easyreporting class object}

\item{optionList}{a list of options}
}
\value{
an easyreporting class object
}
\description{
internal function for appending to the report the initial
options setup
}
\keyword{internal}
