% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdSetGlobalOpts}
\alias{mkdSetGlobalOpts}
\title{mkdSetGlobalOpts-generic}
\usage{
mkdSetGlobalOpts(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
mkdSetGlobalOpts-generic
}
\keyword{internal}
