% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mkdSourceFiles,easyreporting-method}
\alias{mkdSourceFiles,easyreporting-method}
\title{mkdSourceFiles}
\usage{
\S4method{mkdSourceFiles}{easyreporting}(object, ...)
}
\arguments{
\item{object}{an easyreporting class object}

\item{...}{a list of files to source with path}
}
\value{
none
}
\description{
includes a list of source files inside the rmarkdown
}
\keyword{internal}
