% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdVariableAssignment}
\alias{mkdVariableAssignment}
\title{mkdVariableAssignment-generic}
\usage{
mkdVariableAssignment(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
mkdVariableAssignment-generic
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
## leaving the default options to the code chunk
mkdCodeChunkSt(rd)
## adding a variable assignement
variable <- 1
mkdVariableAssignment(rd, "variable", "variable", show=TRUE)
mkdCodeChunkEnd(rd)
}
