% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setBibliography,easyreporting-method}
\alias{setBibliography,easyreporting-method}
\title{setBibliography}
\usage{
\S4method{setBibliography}{easyreporting}(object, bibfile = NULL)
}
\arguments{
\item{object}{an easyreporting class object}

\item{bibfile}{a string with the name of the bib file}
}
\value{
none
}
\description{
add a bibfile name to the object that will be reflected into the
report as a bibliography section
}
\examples{
\dontrun{
# TBD
}
}
\keyword{internal}
