% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{setBibliography}
\alias{setBibliography}
\title{setBibliography}
\usage{
setBibliography(object, bibfile = NULL)
}
\arguments{
\item{object}{an easyreporting class object}

\item{bibfile}{a path string with the name of the bib file}
}
\value{
none
}
\description{
add a bibfile name to the object that will be reflected into the
report as a bibliography section
}
\examples{
\dontrun{
TBD
}
}
