% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{setOptionsList,easyreporting-method}
\alias{setOptionsList,easyreporting-method}
\title{setOptionsList}
\usage{
\S4method{setOptionsList}{easyreporting}(
  object,
  cacheFlag = TRUE,
  evalFlag = TRUE,
  echoFlag = TRUE,
  warningFlag = FALSE,
  showMessages = FALSE,
  includeFlag = TRUE,
  collapseFlag = FALSE,
  purlFlag = TRUE,
  errorFlag = TRUE,
  messageFlag = TRUE,
  highlightFlag = TRUE,
  promptFlag = FALSE,
  stripWhiteFlag = TRUE,
  tidyFlag = FALSE
)
}
\arguments{
\item{object}{an easyreporting class object}

\item{cacheFlag}{boolean for caching chunk data (default TRUE)}

\item{evalFlag}{boolean for evaluating the code chunk in the compiled version
(default TRUE)}

\item{echoFlag}{boolean for showing the code chunk (default TRUE)}

\item{warningFlag}{boolean for showing the chunk warnings (default FALSE)}

\item{showMessages}{boolean for showing the chunk warnings in compiled}

\item{includeFlag}{boolean for including the code chunk in the compiled
version (default TRUE)}
}
\value{
none
}
\description{
set an optionList to the class
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))

## setting default option
setOptionsList(rd)

## modifying only some options
rd <- setOptionsList(rd, warningFlag=TRUE, 
        showMessages=TRUE, includeFlag=TRUE)

}
