% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{setOptionsList}
\alias{setOptionsList}
\title{setOptionsList-generic}
\usage{
setOptionsList(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
setOptionsList-generic
}
\keyword{internal}
