% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{compile}
\alias{compile}
\title{compile}
\usage{
compile(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
none
}
\description{
prints the sessionInfo and compiles the produced rmarkdown file
into an HTML report.
}
\examples{
\dontrun{
rd <- easyreporting(filenamePath="./project_report", title="example_report",
                        author=c("It's me"))
compile(rd)
}
}
