% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{compile}
\alias{compile}
\title{compile-generic}
\usage{
compile(object)
}
\arguments{
\item{object}{a general object}
}
\value{
none
}
\description{
compiles the rmarkdown file
}
\examples{
rd <- easyreporting(filenamePath="./project_report", title="example_report",
                        author=c("It's me"))
compile(rd)
}
