% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingC.R
\docType{class}
\name{easyreporting-class}
\alias{easyreporting-class}
\alias{easyreporting}
\title{easyreporting}
\usage{
easyreporting(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
}
\arguments{
\item{filenamepath}{the path with the name of the rmarkdown.}

\item{title}{the title fo the report.}

\item{author}{the author of the report.}
}
\value{
an easyreporting class instance
}
\description{
An R6 class for managing rmarkdown report.
Each instance describes an rmarkdown file.
}
\examples{

}
