% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{getOptionsList}
\alias{getOptionsList}
\title{getOptionsList}
\usage{
getOptionsList(object)
}
\arguments{
\item{object}{a general object}
}
\value{
a list of options
}
\description{
returns the optionList from the easyreporting class
}
\examples{
rd <- easyreporting(filenamePath="./project_report", title="example_report",
                        author=c("It's me"))
optList <- getOptionsList(rd)

}
