% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{getOptionsList}
\alias{getOptionsList}
\title{getOptionsList-generic}
\usage{
getOptionsList(object)
}
\arguments{
\item{object}{a general object}
}
\value{
none
}
\description{
getOptionsList-generic
}
\examples{
rd <- easyreporting(filenamePath="./project_report", title="example_report",
                        author=c("It's me"))
optList <- getOptionsList(rd)

}
