% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getOptionsList}
\alias{getOptionsList}
\title{getOptionsList}
\usage{
getOptionsList(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
a list of options
}
\description{
returns the optionList from the easyreporting class 
(see the makeOptionList function for more details).
}
\examples{
\dontrun{
rd <- easyreporting(filenamePath="./project_report", title="example_report",
                        author=c("It's me"))
optList <- getOptionsList(rd)
}
}
