% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6MarkdownMethods.R
\name{getReportFilename}
\alias{getReportFilename}
\title{getReportFilename}
\usage{
getReportFilename()
}
\value{
a string of report file name with path
}
\description{
returns the report filename with path
}
\examples{
rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
rep <- rd$getReportFilename()
}
