% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getReportFilename}
\alias{getReportFilename}
\title{getReportFilename}
\usage{
getReportFilename(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
a string of report file name with path
}
\description{
returns the filename with its path of the report generated by 
easyreporting
}
\examples{
\dontrun{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
(rep <- getReportFilename(rd))
}
}
