% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{initReportFilename,easyreporting-method}
\alias{initReportFilename,easyreporting-method}
\title{initReportFilename}
\usage{
\S4method{initReportFilename}{easyreporting}(
  object,
  filenamepath = NULL,
  title = NULL,
  author = NULL,
  optionList = NULL
)
}
\arguments{
\item{object}{an easyreporting class object}

\item{author}{the name of the report author}

\item{filenamePath}{the name of the report with or without the path}

\item{mainTitle}{the title of the report}

\item{documentType}{type of report final document. (html is default)}
}
\value{
an easyreporting class object
}
\description{
internal funtion usefull to init the report for the first time.
}
\keyword{internal}
