% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{initReportFilename,easyreporting-method}
\alias{initReportFilename,easyreporting-method}
\title{initReportFilename}
\usage{
\S4method{initReportFilename}{easyreporting}(
  object,
  filenamepath = NULL,
  title = NULL,
  author = NULL,
  optionList = NULL,
  bibfile = NULL,
  documentType = "rmarkdown::html_document"
)
}
\arguments{
\item{object}{an easyreporting class object}

\item{filenamepath}{the name of the report with or without the path}

\item{author}{the name of the report author as a \link[link]{person} format 
(where expected \code{comment} named fields are \code{ORCID}, \code{url}, 
\code{affiliation}, \code{affiliation_url}).
In case author is a \code{character}, it will be reported as it is.}

\item{bibfile}{a bibfile for bibliography.}

\item{documentType}{type of report final document, if author(s) is a \code{person} 
it will be \code{distill::distill_article} (\code{rmarkdown::html_document} default)}

\item{mainTitle}{the title of the report}

\item{metadataList}{}
}
\value{
an easyreporting class object
}
\description{
internal funtion usefull to init the report for the first time.
}
\seealso{
\link[utils]{person}
}
\keyword{internal}
