% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6MarkdownMethods.R
\name{initReportFilename}
\alias{initReportFilename}
\title{initReportFilename}
\usage{
initReportFilename(filenamepath = NULL, mainTitle = NULL,
  author = NULL, documentType = "html", optionsList = NULL)
}
\arguments{
\item{filenamepath}{the name of the report with or without the path}

\item{mainTitle}{the title of the report}

\item{author}{the name of the report author}

\item{documentType}{type of report final document. (html is default)}
}
\value{
none
}
\description{
internal funtion usefull to init the report for the first time.
}
\keyword{internal}
