% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{initReportFilename}
\alias{initReportFilename}
\title{initReportFilename}
\usage{
initReportFilename(
  object,
  filenamepath = NULL,
  title = NULL,
  author = NULL,
  optionList = NULL,
  bibfile = NULL,
  documentType = NULL,
  resources = NULL
)
}
\arguments{
\item{object}{an easyreporting class object}

\item{filenamepath}{the name of the report with or without the path}

\item{author}{the name of the report author}

\item{documentType}{type of report final document. (html is default)}

\item{mainTitle}{the title of the report}
}
\value{
an easyreporting class object
}
\description{
initReportFilename
}
\keyword{internal}
