% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdCodeChunkCommented,easyreporting-method}
\alias{mkdCodeChunkCommented,easyreporting-method}
\title{mkdCodeChunkCommented}
\usage{
\S4method{mkdCodeChunkCommented}{easyreporting}(
  object,
  commentMsg = NULL,
  codeMsg,
  optionsList = getOptionsList(object),
  sourceFilesList = NULL
)
}
\arguments{
\item{object}{an easyreporting class object}

\item{commentMsg}{a string with the natural language comment for the chunk.}

\item{codeMsg}{a string within the code.}

\item{optionsList}{a list of options (default is the class options).}

\item{sourceFilesList}{a optional list of files to source inside the chunk.}
}
\value{
none
}
\description{
it creates a complete code chunk, adding a natural language
comment before of it.
}
\examples{
rd <- easyreporting(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
optList <- makeOptionsList(includeFlag=TRUE, cacheFlag=TRUE)
mkdCodeChunkCommented(rd,
                commentMsg="This is the comment of the following code chunk",
                codeMsg="a <- 1\\nb <- 2\\n(c <- a+b)\\n", optionsList=optList,
                sourceFilesList=NULL)

}
