% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdCodeChunkCommented}
\alias{mkdCodeChunkCommented}
\title{mkdCodeChunkCommented-generic}
\usage{
mkdCodeChunkCommented(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
mkdCodeChunkCommented-generic
}
\keyword{internal}
