% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6MarkdownMethods.R
\name{mkdCodeChunkCommented}
\alias{mkdCodeChunkCommented}
\title{mkdCodeChunkCommented}
\usage{
mkdCodeChunkCommented(commentMsg = NULL, codeMsg,
  optionsList = self$getOptionsList(), source.files.list = NULL)
}
\arguments{
\item{commentMsg}{a string with the natural language comment for the chunk.}

\item{codeMsg}{a string within the code.}

\item{optionsList}{a list of options (default is the class options).}

\item{source.files.list}{a optional list of files to source inside the chunk.}
}
\value{
none
}
\description{
it creates a complete code chunk, adding a natural language
comment before of it.
}
\examples{
rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
rd$mkdCodeChunkCommented(
                commentMsg="This is the comment of the following code chunk",
                codeMsg="a <- 1\\nb <- 2\\n(c <- a+b)\\n", optionsList=optList,
                source.files.list=NULL)
}
