% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdCodeChunkComplete,easyreporting-method}
\alias{mkdCodeChunkComplete,easyreporting-method}
\title{mkdCodeChunkComplete}
\usage{
\S4method{mkdCodeChunkComplete}{easyreporting}(
  object,
  message,
  optionList = getOptionsList(object),
  sourceFilesList = NULL
)
}
\arguments{
\item{object}{an easyreporting class object}

\item{message}{a string containing a function call or the entire code chunk
to trace.}

\item{optionList}{a list of options.}

\item{sourceFilesList}{a list of files to source.}
}
\value{
none
}
\description{
it creates a complete code chunk.
}
\examples{
 rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
mkdCodeChunkComplete(rd, message="a <- 1\nb <- 2\nc <- a+b\n print(c)")
}
