% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdCodeChunkComplete}
\alias{mkdCodeChunkComplete}
\title{mkdCodeChunkComplete}
\usage{
mkdCodeChunkComplete(
  message,
  optionsList = self$getOptionsList(),
  sourceFilesList = NULL
)
}
\arguments{
\item{message}{a string containing a function call or the entire code chunk
to trace.}

\item{optionsList}{a list of options.}

\item{sourceFilesList}{a list of files to source.}
}
\value{
none
}
\description{
it creates a complete code chunk.
}
\examples{
 rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
rd$mkdCodeChunkComplete(message="a <- 1\nb <- 2\nc <- a+b\n print(c)")
}
