% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdCodeChunkComplete}
\alias{mkdCodeChunkComplete}
\title{mkdCodeChunkComplete-generic}
\usage{
mkdCodeChunkComplete(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
mkdCodeChunkComplete-generic
}
\keyword{internal}
