% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdCodeChunkComplete}
\alias{mkdCodeChunkComplete}
\title{mkdCodeChunkComplete-generic}
\usage{
mkdCodeChunkComplete(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
mkdCodeChunkComplete-generic
}
\examples{
 rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
mkdCodeChunkComplete(rd, message="a <- 1\nb <- 2\nc <- a+b\n print(c)")
}
