% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdCodeChunkEnd}
\alias{mkdCodeChunkEnd}
\title{mkdCodeChunkEnd}
\usage{
mkdCodeChunkEnd(object)
}
\arguments{
\item{object}{an easyreporting class object}
}
\value{
none
}
\description{
it creates a code chunk end. Always use it after a
mkdCodeChunkSt()
}
\examples{
 rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
mkdCodeChunkSt(rd)
## just leaving empty
mkdCodeChunkEnd(rd)
}
