% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdCodeChunkEnd}
\alias{mkdCodeChunkEnd}
\title{mkdCodeChunkEnd}
\usage{
mkdCodeChunkEnd()
}
\value{
none
}
\description{
it creates a code chunk end. Always use it after a
mkdCodeChunkSt()
}
\examples{
 rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
rd$mkdCodeChunkSt()
## just leaving empty
rd$mkdCodeChunkEnd()
}
