% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdCodeChunkEnd}
\alias{mkdCodeChunkEnd}
\title{mkdCodeChunkEnd-generic}
\usage{
mkdCodeChunkEnd(object)
}
\arguments{
\item{object}{a general object}
}
\value{
none
}
\description{
mkdCodeChunkEnd-generic
}
\keyword{internal}
