% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdCodeChunkSt}
\alias{mkdCodeChunkSt}
\title{mkdCodeChunkSt-generic}
\usage{
mkdCodeChunkSt(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
mkdCodeChunkSt-generic
}
\keyword{internal}
