% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{mkdGeneralMsg}
\alias{mkdGeneralMsg}
\title{mkdGeneralMsg}
\usage{
mkdGeneralMsg(object, message)
}
\arguments{
\item{object}{an easyreporting class object}

\item{message}{the message to append to the report}
}
\value{
none
}
\description{
It appends a general message to the report, useful for 
describing a code chunk or a part of the report.
Useful for adding natural language comments.
}
\examples{
\dontrun{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
mkdGeneralMsg(rd, message="Writing a paragraph to describe my code chunk")
}
}
