% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdGeneralMsg}
\alias{mkdGeneralMsg}
\title{mkdGeneralMsg}
\usage{
mkdGeneralMsg(object, message)
}
\arguments{
\item{object}{an easyreporting class object}

\item{message}{the message to append to the report}
}
\value{
none
}
\description{
mkdGeneralMsg
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
mkdGeneralMsg(rd, message="Writing a paragraph to describe my code chunk")
}
