% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdGeneralMsg}
\alias{mkdGeneralMsg}
\title{mkdGeneralMsg}
\usage{
mkdGeneralMsg(message)
}
\arguments{
\item{message}{the message to append to the report}
}
\value{
none
}
\description{
It appends a general message to the report.
Useful for adding natural language comments.
}
\examples{
rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
rd$mkdGeneralMsg("Writing a paragraph to describe my code chunk")
}
