% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdGeneralTitledMsg}
\alias{mkdGeneralTitledMsg}
\title{mkdGeneralTitledMsg}
\usage{
mkdGeneralTitledMsg(object, title = NULL, level = 1, message)
}
\arguments{
\item{object}{an easyreporting class object}

\item{level}{the level of the title}

\item{message}{the message to append to the report}
}
\value{
none
}
\description{
It appends a a titled section followed by a general message to 
the report. Useful for adding natural language comments.
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))
mkdGeneralMsg(rd, message="Writing a paragraph to describe my code chunk")
}
