% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{mkdSetGlobalOpts}
\alias{mkdSetGlobalOpts}
\title{mkdSetGlobalOpts}
\usage{
mkdSetGlobalOpts(object, optionList = list())
}
\arguments{
\item{object}{an easyreporting class object}

\item{optionList}{a list of options}
}
\value{
an easyreporting class object
}
\description{
mkdSetGlobalOpts
}
\keyword{internal}
