% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6MarkdownMethods.R
\name{mkdSetGlobalOpts}
\alias{mkdSetGlobalOpts}
\title{mkdSetGlobalOpts}
\usage{
mkdSetGlobalOpts(optionList)
}
\arguments{
\item{optionsList}{a list of options}
}
\value{
none
}
\description{
internal function for appending to the report the initial
options setup
}
\keyword{internal}
