% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdSourceFiles}
\alias{mkdSourceFiles}
\title{mkdSourceFiles-generic}
\usage{
mkdSourceFiles(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
mkdSourceFiles-generic
}
\keyword{internal}
