% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdSourceFiles}
\alias{mkdSourceFiles}
\title{mkdSourceFiles}
\usage{
mkdSourceFiles(...)
}
\arguments{
\item{...}{a list of files to source}
}
\value{
none
}
\description{
includes a list of source files inside the rmarkdown
}
\keyword{internal}
