% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingM.R
\name{mkdTitle}
\alias{mkdTitle}
\title{mkdTitle}
\usage{
mkdTitle(title, level = 1)
}
\arguments{
\item{title}{a string within the title.}

\item{level}{a numeric from 1 to 6 (default is 1).}
}
\value{
none
}
\description{
Inserts an rmarkdown title inside the report.
}
\examples{
rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))

rd$mkdTitle("First Level Title")
rd$mkdTitle("Second Level Title", level=2)

}
