% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdTitle}
\alias{mkdTitle}
\title{mkdTitle}
\usage{
mkdTitle(object, title, level = 1)
}
\arguments{
\item{object}{an easyreporting class object}

\item{title}{a string within the title.}

\item{level}{a numeric from 1 to 6 (default is 1).}
}
\value{
none
}
\description{
Inserts an rmarkdown title inside the report, useful for create
distinct sections inside the final resport.
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))

mkdTitle(rd, "First Level Title")
mkdTitle(rd, "Sub-Title", level=2)

}
