% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6MarkdownMethods.R
\name{mkdVariableAssignment}
\alias{mkdVariableAssignment}
\title{mkdVariableAssignment}
\usage{
mkdVariableAssignment(variable.name, variable.object.name, show = FALSE)
}
\arguments{
\item{variable.name}{a string indicating the name of the variabe to store in
the report. (This can be changed here, but further uses of the variable needs
to take into account the variable name change).}

\item{variable.object.name}{the name of the already existing variable. (This
cannot be canged.)}

\item{show}{a boolean indicating if to show the message before writing it
into the rmardown file.}
}
\value{
none
}
\description{
it includes a variable assignment in the report.
NB: a call to the "mkdCodeChunkSt" has to be done before using it.
}
\examples{
rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))
## leaving the default options to the code chunk
rd$mkdCodeChunkSt()
## adding a variable assignement
variable <- 1
rd$mkdVariableAssignment("variable", "variable", show=TRUE)
rd$mkdCodeChunkEnd()

}
