% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{mkdVariableAssignment}
\alias{mkdVariableAssignment}
\title{mkdVariableAssignment-generic}
\usage{
mkdVariableAssignment(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
mkdVariableAssignment-generic
}
\keyword{internal}
