% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6MarkdownMethods.R
\name{setOptionsList}
\alias{setOptionsList}
\title{setOptionsList}
\usage{
setOptionsList(cacheFlag = TRUE, evalFlag = TRUE, echoFlag = TRUE,
  warningFlag = FALSE, showMessages = FALSE, includeFlag = TRUE)
}
\arguments{
\item{cacheFlag}{boolean for caching chunk data (default TRUE)}

\item{evalFlag}{boolean for evaluating the code chunk in the compiled version
(default TRUE)}

\item{echoFlag}{boolean for showing the code chunk (default TRUE)}

\item{warningFlag}{boolean for showing the chunk warnings (default FALSE)}

\item{showMessages}{boolean for showing the chunk warnings in compiled}

\item{includeFlag}{boolean for including the code chunk in the compiled
version (default TRUE)}
}
\value{
none
}
\description{
set an optionList to the rmarkdown R6 class
}
\examples{
rd <- easyreporting$new(filenamepath="./project_report",
                        title="example_report", author=c("It's me"))

## setting default option
rd$setOptionsList()

## modifying only some options
rd$setOptionsList(warningFlag=TRUE, showMessages=TRUE, includeFlag=TRUE)

}
