% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreportingG.R
\name{setOptionsList}
\alias{setOptionsList}
\title{setOptionsList-generic}
\usage{
setOptionsList(object, ...)
}
\arguments{
\item{object}{a general object}

\item{...}{whatever}
}
\value{
none
}
\description{
setOptionsList-generic
}
\examples{
rd <- easyreporting(filenamePath="./project_report",
                        title="example_report", author=c("It's me"))

## setting default option
setOptionsList(rd)

## modifying only some options
rd <- setOptionsList(rd, warningFlag=TRUE, 
        showMessages=TRUE, includeFlag=TRUE)

}
